backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
joblib
motor
pymongo
python-dotenv